// models/structsButtonMessageWithMediaPayload.ts
/**
 * WhatsAPI Go
 * The V2 of WhatsAPI Go
 *
 * The version of the OpenAPI document: 2.0
 * Contact: manjit@sponsorbook.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { StructsFileUpload } from './structsFileUpload';
import { StructsReplyButton } from './structsReplyButton';

export class StructsButtonMessageWithMediaPayload {
    'buttons'?: Array<StructsReplyButton>;
    'footerText'?: string;
    'mediaData'?: StructsFileUpload;
    'text'?: string;
    'to'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "buttons",
            "baseName": "buttons",
            "type": "Array<StructsReplyButton>"
        },
        {
            "name": "footerText",
            "baseName": "footer_text",
            "type": "string"
        },
        {
            "name": "mediaData",
            "baseName": "media_data",
            "type": "StructsFileUpload"
        },
        {
            "name": "text",
            "baseName": "text",
            "type": "string"
        },
        {
            "name": "to",
            "baseName": "to",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return StructsButtonMessageWithMediaPayload.attributeTypeMap;
    }
}