// models/structsGroupCreatePayload.ts
/**
 * WhatsAPI Go
 * The V2 of WhatsAPI Go
 *
 * The version of the OpenAPI document: 2.0
 * Contact: manjit@sponsorbook.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class StructsGroupCreatePayload {
    'groupName'?: string;
    'participants'?: Array<string>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "groupName",
            "baseName": "group_name",
            "type": "string"
        },
        {
            "name": "participants",
            "baseName": "participants",
            "type": "Array<string>"
        }    ];

    static getAttributeTypeMap() {
        return StructsGroupCreatePayload.attributeTypeMap;
    }
}