// models/structsListMessagePayload.ts
/**
 * WhatsAPI Go
 * The V2 of WhatsAPI Go
 *
 * The version of the OpenAPI document: 2.0
 * Contact: manjit@sponsorbook.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { StructsListSection } from './structsListSection';

export class StructsListMessagePayload {
    'buttonText'?: string;
    'description'?: string;
    'multiSelect'?: boolean;
    'sections': Array<StructsListSection>;
    'text'?: string;
    'title'?: string;
    'to': string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "buttonText",
            "baseName": "button_text",
            "type": "string"
        },
        {
            "name": "description",
            "baseName": "description",
            "type": "string"
        },
        {
            "name": "multiSelect",
            "baseName": "multi_select",
            "type": "boolean"
        },
        {
            "name": "sections",
            "baseName": "sections",
            "type": "Array<StructsListSection>"
        },
        {
            "name": "text",
            "baseName": "text",
            "type": "string"
        },
        {
            "name": "title",
            "baseName": "title",
            "type": "string"
        },
        {
            "name": "to",
            "baseName": "to",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return StructsListMessagePayload.attributeTypeMap;
    }
}